Add default editor; use $VISUAL before $EDITOR
Treat the editor value, whether it was given via the command line
option or read from an environment variable, as a command + its
arguments rather than just a command. This means that if the path
to your editor contains spaces you now need to double- or
curly-quote it. E.g.,
$ gpgedit -editor '"/a/path/with spaces/e"'

Remove the unnecessary fileutil dependency.

Style.

v0.1.2